requirements.txt
jira
python-dotenv